pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unify-llm"
version = "0.1.0"
description = "A unified interface for calling various Large Language Model APIs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "UnifyLLM Team"}
]
keywords = ["llm", "openai", "anthropic", "gemini", "ollama", "unified-api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "tenacity>=8.0.0",
    "typing-extensions>=4.5.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rootutils>=1.0.7",
    "pydantic-settings>=2.11.0",
    # n8n-style agent dependencies
    "croniter>=2.0.0",  # For schedule triggers
    "fastapi>=0.110.0",  # For webhook server
    "uvicorn>=0.27.0",  # For webhook server
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/unifyllm/unifyllm"
Documentation = "https://unifyllm.readthedocs.io"
Repository = "https://github.com/unifyllm/unifyllm"

[tool.setuptools.packages.find]
include = ["unify_llm*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=unify_llm --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests that call real LLM APIs (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
